+++ src/pathlib_next/protocols/__init__.py
@@ -1,2 +1,3 @@
 from .fs import *
 from .io import *
+from .checksum import *